Replace minimum and maximum heap size declarations with single declaration of `elasticsearch_heap_size`.
This replacement is to better line up with both the given comments in the jvm.options.j2 template stating that maximum heap size must equal minimum heap size, as well as to match the official elastic repository. Variability between the two can result in elasticsearch either not running or instability.